Consolidate battery controllers and formatting.
