+++ whoc/controllers/__init__.py
@@ -1,5 +1,4 @@
-from whoc.controllers.battery_controller import BatteryController
-from whoc.controllers.battery_passthrough_controller import BatteryPassthroughController
+from whoc.controllers.battery_controller import BatteryController, BatteryPassthroughController
 from whoc.controllers.hybrid_supervisory_controller import HybridSupervisoryControllerBaseline
 from whoc.controllers.lookup_based_wake_steering_controller import LookupBasedWakeSteeringController
 from whoc.controllers.solar_passthrough_controller import SolarPassthroughController
