+++ whoc/controllers/battery_controller.py
@@ -60,3 +60,15 @@ class BatteryController(ControllerBase):
         # adding integral action. Not sure that's what I want? Only if I add it 
         # to the existing reference?
         self.controls_dict["power_setpoint"] = current_power + u
+
+
+class BatteryPassthroughController(ControllerBase):
+    """
+    Simply passes power reference down to (single) battery.
+    """
+    def __init__(self, interface, input_dict, verbose=True):
+        super().__init__(interface, verbose)
+
+    def compute_controls(self):
+        reference_power = self.measurements_dict["battery_power_reference"]
+        self.controls_dict["power_setpoint"] = reference_power
