+++ whoc/controllers/wind_farm_power_tracking_controller.py
@@ -50,6 +50,7 @@ class WindFarmPowerDistributingController(ControllerBase):
 
         return None
 
+
 class WindFarmPowerTrackingController(WindFarmPowerDistributingController):
     """
     Based on controller developed under A2e2g project. Proportional control only---
